requirements.txt
streamlit>=1.28.0
PyGithub>=2.1.1
gTTS>=2.4.0
moviepy>=1.0.3
manim>=0.18.0
markdown>=3.5.1
requests>=2.31.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
tree-sitter>=0.20.4
ast-comments>=1.0.0 